refactor
moved |[wentry| routines out of |pwfile| and into |pwentry|
