+++ include/pwentry.h
@@ -75,6 +75,12 @@ typedef PWENTRY		pwentry ;
 
 EXTERNC_begin
 
+extern int	pwentry_start(pwentry *) noex ;
+extern int	pwentry_finish(pwentry *) noex ;
+extern int	pwentry_fieldpw(pwentry *,int,cchar *,int) noex ;
+extern int	pwentry_mkextras(pwentry *) noex ;
+extern int	pwentry_mkcopy(pwentry *,pwentry *,char *,int) noex ;
+
 extern int	pwentry_bufsize(void) noex ;
 extern int	getpwentrybufsize(void) noex ;
